+++ docs/change_log.md
@@ -1,1 +1,9 @@
 # Change Log
+
+## 0.2
+
+Added the following features:
+
+1. Download of configuration file.
+2. Download of balance data
+3. Download of income and expense data